requirements.txt
numpy==2.2.0
opencv-python==4.11.0.86
pyserial==3.5
typing_extensions==4.12.2
